feat: TMS-30557: add regular links parsing; refactor;

+++ Migrators/AllureExporter/Models/AllureAttachment.cs
@@ -5,7 +5,7 @@ namespace AllureExporter.Models;
 public class AllureAttachment
 {
     [JsonPropertyName("id")]
-    public int Id { get; set; }
+    public long Id { get; set; }
     [JsonPropertyName("name")]
     public string Name { get; set; } = string.Empty;
 
@@ -20,5 +20,5 @@ public class AllureAttachmentContent
     public List<AllureAttachment> Content { get; set; }  = new ();
 
     [JsonPropertyName("totalPages")]
-    public int TotalPages { get; set; }
+    public long TotalPages { get; set; }
 }
