+++ Migrators/AllureExporter/Services/ISectionService.cs
@@ -4,5 +4,5 @@ namespace AllureExporter.Services;
 
 public interface ISectionService
 {
-    Task<SectionInfo> ConvertSection(int projectId);
+    Task<SectionInfo> ConvertSection(long projectId);
 }
