+++ Migrators/AllureExporter/Services/IStepService.cs
@@ -4,6 +4,6 @@ namespace AllureExporter.Services;
 
 public interface IStepService
 {
-    Task<List<Step>> ConvertStepsForTestCase(int testCaseId, Dictionary<string, Guid> sharedStepMap);
-    Task<List<Step>> ConvertStepsForSharedStep(int sharedStepId);
+    Task<List<Step>> ConvertStepsForTestCase(long testCaseId, Dictionary<string, Guid> sharedStepMap);
+    Task<List<Step>> ConvertStepsForSharedStep(long sharedStepId);
 }
